+++ lib/widgets/canvas.dart
@@ -1,1 +1,1 @@
-
+part of '../main.dart';
