+++ lib/widgets/creator_widget.dart
@@ -1,15 +1,4 @@
-import 'dart:math';
-
-import 'package:editicert/logic/component_service.dart';
-import 'package:editicert/models/component.dart';
-import 'package:editicert/state/canvas_events_cubit.dart';
-import 'package:editicert/state/canvas_transform_cubit.dart';
-import 'package:editicert/state/keys_cubit.dart';
-import 'package:editicert/state/tool_cubit.dart';
-import 'package:editicert/utils.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
+part of '../main.dart';
 
 class CreatorWidget extends StatefulWidget {
   const CreatorWidget({super.key});
